Separate bowling.rb for each class
